CEntity, CReferences, CTrafficLights, CMotionBlurStreaks, several smallish other classes and random functions (#39)
* CEntity::Add, CEntity::Add_, CEntity::Remove

* Name fix

* SetIsStatic, DeleteRwObject, ProcessControl, ProcessCollision, ProcessShift

* TestCollision, Teleport, SpecialEntityPreCollisionStuff, SpecialEntityCalcCollisionSteps

* CEntity::PreRender, fixes, all ModelIndices

Some small missing classes added

* Render, SetupLighting, RemoveLighting, FlagToDestroyWhenNextProcessed

* All virtual methods working

* UpdateRwFrame, UpdateRpHAnim, HasPreRenderEffects

* CMotionBlurStreaks, CRegisteredMotionBlurStreak

* CTrafficLights

* CTrafficLights working

* Enum rename missing file

* Shitload of fixes and reversed functions

* CEntity::CreateEffects() working, random stuff arounds that

* PtrLists cleanups

* CPhysical cleanups

* Shitload of methods and fixes once again, trying to find the crash cause

* FUrther work

* CAutomobile::PlaceOnRoadProperly()

* Fixes

* BaseModelInfo VTable cleanup

* CalculateBBProjection, UpdateAnim, IsVisible

* VTable cleanup

* CReferences

* CleanUpOldReference, ResolveReferences, PruneReferences

* ModifyMatrixForBannerInWind

* COcclusion stuff

* RegisterReference

* ProcessLightsForEntity first version

* CEntity::ProcessLightsForEntity working seemingly properly

* RemoveEscalatorsForEntity

* CEntity::IsEntityOccluded

* Occlussion stuff

* Occlussion working properly

* Warning cleanups and stuff

* Cleanups

* Cleanups

* Constructor, virtual destructor

* Removed misleading comments

* fix compile-time errors

+++ source/game_sa/Entity/Vehicle/CPlane.h
@@ -49,7 +49,7 @@ public:
     float field_998;
     int field_99C;
     int field_9A0;
-    int field_9A4;
+    float field_9A4;
     float field_9A8;
     float field_9AC;
     float field_9B0;
